fix(run): Changed the run method to get input from validation mw
Changed the run method to get the post data from the formatted property
set by the validation middleware instead of the request body.  This
will make it easier to do transforms on properties not supported in
json, like date and mongoid if we go with mongodb for a data store.